create readme.md
this is the first download of my game added to it:1-music,2-pictures of the player and the enemy, 3-the movement of the player along the arrows and the movement of the enemy will begin after you press Enter,4-there is also a background,5-you can collect coins. That's it!